refactor: removing nested and useless code treatments

+++ src/main/java/com/school/management/service/CourseService.java
@@ -2,11 +2,15 @@ package com.school.management.service;
 
 import com.school.management.model.Course;
 import com.school.management.model.dto.CourseDTO;
+import com.school.management.model.dto.StudentCourseDTO;
+import com.school.management.model.dto.StudentDTO;
 import com.school.management.repository.CourseRepository;
 import com.school.management.repository.StudentCourseRepository;
+import com.school.management.repository.StudentRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.sql.Timestamp;
 import java.time.Instant;
@@ -21,48 +25,43 @@ public class CourseService {
 
     private final StudentCourseRepository studentCourseRepository;
 
-    public List<CourseDTO> createCoursesList(List<CourseDTO> coursesDTO){
+    private final StudentRepository studentRepository;
 
+    public List<CourseDTO> createCoursesList(List<CourseDTO> coursesDTO) {
         Timestamp ts = Timestamp.from(Instant.now());
-        List<Course> courses = courseRepository.saveAll(coursesDTO.stream()
-                .filter(c -> c.getName() != null && !c.getName().isBlank())
-                .map(courseDTO -> new Course(courseDTO.getName(),
-                        ts,
-                        ts))
-                .collect(Collectors.toList()));
-
-        return courses.stream()
-                .map(student -> new CourseDTO(
-                        student.getName(),
-                        student.getCreatedAt(),
-                        student.getUpdatedAt()))
-                .collect(Collectors.toList());
+        List<Course> courses = courseRepository.saveAll(coursesDTO.stream().filter(c -> c.getName() != null && !c.getName().isBlank()).map(courseDTO -> new Course(courseDTO.getName(), ts, ts)).collect(Collectors.toList()));
+        return courses.stream().map(student -> new CourseDTO(student.getName(), student.getCreatedAt(), student.getUpdatedAt())).collect(Collectors.toList());
     }
 
     public CourseDTO findById(Long id) {
-        return courseRepository.findById(id).map(course -> new CourseDTO(course.getId(), course.getName(), course.getCreatedAt(), course.getUpdatedAt())).orElse(null);
+        return courseRepository.findById(id).map(CourseDTO::new).orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Course not found."));
     }
 
-    public void deleteById(Long id, Boolean orElse) {
+    public void deleteById(Long id, Boolean confirmDelete) {
+        if (!confirmDelete) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    "To delete the course and student-courses relationships, inform confirm-deletion=true as a query param.");
+        }
+        courseRepository.findById(id).orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Course not found."));
+
         courseRepository.deleteById(id);
     }
 
     public List<CourseDTO> findCourses(Boolean without_students) {
-        if(without_students){
-            return this.courseRepository.findAllCoursesWithoutStudents().stream().map(course -> new CourseDTO(course.getId(), course.getName(), course.getCreatedAt(), course.getUpdatedAt())).collect(Collectors.toList());
-        }
-        return this.courseRepository.findAllCoursesWithStudents().stream().map(course -> new CourseDTO(course.getId(), course.getName(), course.getCreatedAt(), course.getUpdatedAt())).collect(Collectors.toList());
-
+        List<Course> courses = without_students ? courseRepository.findAllCoursesWithoutStudents() : courseRepository.findAllCoursesWithStudents();
+        return courses.stream().map(CourseDTO::new).collect(Collectors.toList());
     }
 
     public void enrollStudents(Long id, List<Long> studentIds) {
-        studentIds.stream().forEach(studentId -> {
-            this.courseRepository.enrollStudent(id, studentId);
-        } );
+        studentIds.forEach(studentId -> this.courseRepository.enrollStudent(id, studentId));
     }
 
-    public Object getStudentAndCourseRelations() {
-        return studentCourseRepository.findAll().stream();
+    public List<StudentCourseDTO> getStudentAndCourseRelations() {
+        return studentCourseRepository.findAll().stream().map(StudentCourseDTO::new).collect(Collectors.toList());
+
     }
 
     public CourseDTO updateCourse(Long id, CourseDTO courseDTO) {
@@ -72,8 +71,17 @@ public class CourseService {
     }
 
     public void deleteCourses(Boolean confirmDelete) {
-        if (confirmDelete) {
-            courseRepository.deleteAllCoursesAndRelations();
+        if (!confirmDelete) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    "To delete ALL courses and students-courses relationships, inform confirm-deletion=true as a query param.");
         }
+        courseRepository.deleteAllCoursesAndRelations();
+
+    }
+
+    public List<StudentDTO> getStudentsFromCourse(Long id) {
+        courseRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Course not found"));
+        return studentRepository.getStudentsFromCourse(id).stream().map(StudentDTO::new).collect(Collectors.toList());
     }
 }
